Extract class function and test
