+++ packages/base/Grid/src/Grid/Grid.tsx
@@ -13,6 +13,7 @@ import type {
   GridSpacing,
   GridWrap,
 } from '../types'
+import { getGridSpacingClassName } from './utils/get-grid-spacing-class-name'
 export interface Props extends BaseProps, HTMLAttributes<HTMLElement> {
   /** Grid content containing Grid.Item */
   children?: ReactNode
@@ -77,26 +78,6 @@ const justifyContentClassNamesMapping: { [K in GridJustification]: string } = {
   'space-between': 'justify-between',
   'space-around': 'justify-around',
   'space-evenly': 'justify-evenly',
-}
-
-const gridWidthAndMarginMapping: { [K in GridSpacing]: string } = {
-  '0': 'width-calc-100pct-0px -m-[0px]',
-  '8': 'width-calc-100pct-8px -m-[4px]',
-  '16': 'width-calc-100pct-16px -m-[8px]',
-  '24': 'width-calc-100pct-24px -m-[12px]',
-  '32': 'width-calc-100pct-32px -m-[16px]',
-  '64': 'width-calc-100pct-64px -m-[32px]',
-  '72': 'width-calc-100pct-72px -m-[36px]',
-  '80': 'width-calc-100pct-80px -m-[40px]',
-}
-
-const getGridSpacingClassName = (spacing: Props['spacing']) => {
-  if (!spacing) {
-    return ''
-  }
-
-  // Negative margin has half of the spacing value to properly handle space on sides of grid
-  return gridWidthAndMarginMapping[spacing]
 }
 
 export const Grid = forwardRef<HTMLDivElement, Props>(function Grid(
